# public/swagger.yaml
swagger: '2.0'
info:
  version: "1.3.0"
  title: ZXInfo API
  description: |
    Welcome to the world's (probaly) greatest Sinclair API!.

    This ZXDB based API is a tool to help developers everywhere create amazing and incredible web sites and apps using data from the 40-plus years area of Sinclair home computers (and clones).
    
    **Ready** to get started?
    
    This is the short version of what you need to do:
    * Be a good API citizen, read and understand how to use the API
    * Keep in touch: tell us about what you're building and talk to other developers
    * Build cool stuff
    
    The ZXInfo API is also the backend for the web site https://zxinfo.dk
    
    Remember that the ZXInfo API suite and portal is an evolving project, developed by dedicated people in their spare time. To the extent that it is possible, we will try to limit any modifications which cause backwards incompatible changes to applications.
    
    Most communication will occur on the community forum on https://spectrumcomputing.co.uk/forums/
    
  contact:
    name: ZXInfo API development
    email: zxinfo_dev@kolbeck.dk
host: api.zxinfo.dk
basePath: /api/zxinfo
...
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /games/{game-id}:
    get:
      tags:
      - "zxinfo"
      summary: "Fetches a single entry by id"
      description: |
        This method fetches a single entry resource. It is the canonical URI for any entry resource provided by the API. Can be shared on e.g. facebook for a direct link to an entry
        
      operationId: "getGameById"
      produces:
      - "application/json"
      parameters:
      - name: "game-id"
        in: path
        description: An entry id (WoS compatible, e.g. 0002259), left padded with 0's if less than 7 digits.
        type: string
        required: true
      - name: mode
        in: query
        description: Specifies if returned JSON should be 'compact' or 'full'
        required: false
        default: "compact"
        type: string
      responses:
        # Response code
        200:
          description: OK
        400:
          description: Bad request. Entry id must be an integer. larger than 0 and max 7 digits.
        404:
          description: An entry with the specified ID was not found.
  /publishers/{publisher-name}/games:
    get:
      summary: "Fetches list of entries filtered by publisher"
      description: |
        Fetches list of entries from a specific publisher, including:
        * publisher.name
        * releases.publisher.name
        * authors.authors.name
        * authors.group
        
      tags:
      - "zxinfo"
      parameters:
        - name: publisher-name
          in: path
          description: Specifies name of publisher to lookup
          required: true
          type: string
        - name: size
          in: query
          description: Specifies the page size for pagination
          required: true
          type: integer
        - name: offset
          in: query
          description: Specifies the page offset for pagination
          required: true
          type: integer
        - name: mode
          in: query
          description: Specifies if returned JSON should be 'compact' or 'full'
          required: false
          default: "compact"
          type: string
        - name: sort
          in: query
          description: Specifies sorting mode. Can be on of title_asc, title_desc, date_asc or date_desc.
          required: false
          default: "date_desc"
          type: string
      responses:
        # Response code
        200:
          description: |
            An array of games. Adds *X-Total-Count* to reponse header.
  /publishers/{publisher-name}/games/{game-title}:
    get:
      summary: Fetches entry by publisher with title
      description: Returns a single game
      tags:
      - "zxinfo"
      parameters:
        - name: publisher-name
          in: path
          description: Publisher or author to query
          required: true
          type: string
        - name: game-title
          in: path
          description: Title to query
          required: true
          type: string
        - name: mode
          in: query
          description: Specifies if returned JSON should be 'compact' or 'full'
          required: false
          default: "compact"
          type: string
      responses:
        # Response code
        200:
          description: "successful operation, JSON document with details"
        404:
          description: An entry with specified publisher and title was not found.
...